requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
Faker==19.2.0
favicon==0.7.0
fonttools==4.41.1
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-secret-manager==2.10.0
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.56.2
grpcio-status==1.56.2
htbuilder==0.6.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
more-itertools==10.0.0
multidict==6.0.4
numpy==1.25.2
openai==0.27.8
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
proto-plus==1.22.3
protobuf==4.23.4
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.0
Pygments==2.15.1
pymdown-extensions==10.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
scipy==1.11.1
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
st-annotated-text==4.0.0
streamlit==1.25.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.0
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.2
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
watchdog==3.0.0
yarl==1.9.2
zipp==3.16.2
